requirements.txt
aniso8601==9.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-HTTPAuth==4.7.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pytz==2023.3
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3

paho-mqtt~=1.6.1